Rename output module
